[중요] Employee salary의 float 부동소수점 문제로 BigDecimal로 변경. 직원 수정시 사번 안바뀌는 것 처리

+++ src/com/konkuk/dto/Employee.java
@@ -2,13 +2,15 @@ package com.konkuk.dto;
 
 import com.konkuk.asset.Langs;
 
+import java.math.BigDecimal;
+
 public class Employee {
     public int id;
     public String name;
-    public float salary;
+    public BigDecimal salary;
     public float residualDayOff;
 
-    public Employee(int id, String name, float salary, float residualDayOff) {
+    public Employee(int id, String name, BigDecimal salary, float residualDayOff) {
         this.id = id;
         this.name = name;
         this.salary = salary;
@@ -23,8 +25,8 @@ public class Employee {
     public void setId(int id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
-    public float getSalary() { return salary; }
-    public void setSalary(float salary) { this.salary = salary; }
+    public BigDecimal getSalary() { return salary; }
+    public void setSalary(BigDecimal salary) { this.salary = salary; }
     public float getResidualDayOff() { return residualDayOff; }
     public void setResidualDayOff(float num) { this.residualDayOff = num; }
 }
